refactor(coolify): keep src/shared for what both processes import
Electron builds the main and renderer bundles separately, so src/shared
exists for modules either side may import — slugify, themes and
version_utils are all used from both. Three Coolify modules were put there
because they were pure, which is why they are easy to test but not why they
would need to be neutral between bundles: the build config and domain
helpers are main-process only, and the insecure-connection warning is
renderer only. Each now sits beside its single caller.

That also settles a layering complaint from review: coolify_build_config
imported its CoolifyBuildConfig type from ipc/utils, a shared module
reaching into main-process code. The import is unremarkable now that the
file is main-process code itself, next to commands.ts which does the same.

Only coolify_scopes stays shared, which is the point of it — the setup
instructions in the renderer and the 403 message in the main process
interpolate one constant so they cannot drift apart again.
